+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v1.2.0 - upcoming
 * verbose option can be given multiple times
-* do full cell calculations 
+* do full cell calculations
+* insert step index into half cycle table 
 
 ## v1.1.7 - 2017-09-30
 * fix linux compilation